requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
openai==0.27.8
pandas==2.0.3
textstat==0.7.3
Werkzeug==2.3.6
Jinja2==3.1.2
requests==2.31.0